+++ app/Controllers/AccountsController.php
@@ -44,7 +44,12 @@ class AccountsController extends Container
     public function index(Request $request, Response $response, array $args) : Response
     {
         if ($this->registry->get('plugins.accounts.settings.index.enabled') === false) {
-            return $response->withRedirect($this->router->pathFor('accounts.login'));
+
+            $theme_template_path  = 'themes/' . $this->registry->get('plugins.site.settings.theme') . '/templates/accounts/templates/no-access.html';
+            $plugin_template_path = 'plugins/accounts/templates/no-access.html';
+            $template_path        = Filesystem::has(PATH['project'] . '/' . $theme_template_path) ? $theme_template_path : $plugin_template_path;
+
+            return $this->twig->render($response, $template_path);
         }
 
         $accounts_list = Filesystem::listContents(PATH['project'] . '/accounts');
@@ -84,6 +89,15 @@ class AccountsController extends Container
      */
     public function login(Request $request, Response $response, array $args) : Response
     {
+        if ($this->registry->get('plugins.accounts.settings.login.enabled') === false) {
+
+            $theme_template_path  = 'themes/' . $this->registry->get('plugins.site.settings.theme') . '/templates/accounts/templates/no-access.html';
+            $plugin_template_path = 'plugins/accounts/templates/no-access.html';
+            $template_path        = Filesystem::has(PATH['project'] . '/' . $theme_template_path) ? $theme_template_path : $plugin_template_path;
+
+            return $this->twig->render($response, $template_path);
+        }
+
         if ($this->acl->isUserLoggedIn()) {
             return $response->withRedirect($this->router->pathFor('accounts.profile', ['email' => $this->acl->getUserLoggedInEmail()]));
         }
@@ -263,6 +277,15 @@ class AccountsController extends Container
      */
     public function resetPassword(Request $request, Response $response, array $args) : Response
     {
+        if ($this->registry->get('plugins.accounts.settings.reset_password.enabled') === false) {
+
+            $theme_template_path  = 'themes/' . $this->registry->get('plugins.site.settings.theme') . '/templates/accounts/templates/no-access.html';
+            $plugin_template_path = 'plugins/accounts/templates/no-access.html';
+            $template_path        = Filesystem::has(PATH['project'] . '/' . $theme_template_path) ? $theme_template_path : $plugin_template_path;
+
+            return $this->twig->render($response, $template_path);
+        }
+
         $theme_template_path  = 'themes/' . $this->registry->get('plugins.site.settings.theme') . '/templates/accounts/templates/registration.html';
         $plugin_template_path = 'plugins/accounts/templates/reset-password.html';
         $template_path        = Filesystem::has(PATH['project'] . '/' . $theme_template_path) ? $theme_template_path : $plugin_template_path;
@@ -376,7 +399,12 @@ class AccountsController extends Container
     public function registration(Request $request, Response $response, array $args) : Response
     {
         if ($this->registry->get('plugins.accounts.settings.registration.enabled') === false) {
-            return $response->withRedirect($this->router->pathFor('accounts.login'));
+
+            $theme_template_path  = 'themes/' . $this->registry->get('plugins.site.settings.theme') . '/templates/accounts/templates/no-access.html';
+            $plugin_template_path = 'plugins/accounts/templates/no-access.html';
+            $template_path        = Filesystem::has(PATH['project'] . '/' . $theme_template_path) ? $theme_template_path : $plugin_template_path;
+
+            return $this->twig->render($response, $template_path);
         }
 
         if ($this->acl->isUserLoggedIn()) {
@@ -511,6 +539,15 @@ class AccountsController extends Container
      */
     public function profile(Request $request, Response $response, array $args) : Response
     {
+        if ($this->registry->get('plugins.accounts.settings.profile.enabled') === false) {
+
+            $theme_template_path  = 'themes/' . $this->registry->get('plugins.site.settings.theme') . '/templates/accounts/templates/no-access.html';
+            $plugin_template_path = 'plugins/accounts/templates/no-access.html';
+            $template_path        = Filesystem::has(PATH['project'] . '/' . $theme_template_path) ? $theme_template_path : $plugin_template_path;
+
+            return $this->twig->render($response, $template_path);
+        }
+
         $email = $args['email'];
 
         // Redirect to accounts index if profile not founded
@@ -547,6 +584,15 @@ class AccountsController extends Container
      */
     public function profileEdit(Request $request, Response $response, array $args) : Response
     {
+        if ($this->registry->get('plugins.accounts.settings.profile_edit.enabled') === false) {
+
+            $theme_template_path  = 'themes/' . $this->registry->get('plugins.site.settings.theme') . '/templates/accounts/templates/no-access.html';
+            $plugin_template_path = 'plugins/accounts/templates/no-access.html';
+            $template_path        = Filesystem::has(PATH['project'] . '/' . $theme_template_path) ? $theme_template_path : $plugin_template_path;
+
+            return $this->twig->render($response, $template_path);
+        }
+
         $email = $args['email'];
 
         // Redirect to accounts index if profile not founded